Move to to_device module
