+++ pyinsteon/handlers/to_device/enter_unlinking_mode.py
@@ -1,7 +1,7 @@
 """Get Device Info command handler."""
 from .direct_command import DirectCommandHandlerBase
-from ..address import Address
-from ..topics import ENTER_UNLINKING_MODE
+from ...address import Address
+from ...topics import ENTER_UNLINKING_MODE
 
 
 class EnterUnlinkingModeCommand(DirectCommandHandlerBase):
